shaders working

+++ libs/opengl/shaders/points.v.glsl
@@ -14,11 +14,11 @@ uniform float vertexPointSize;
 uniform int enableVariablePointSize;  // 0 or 1
 uniform float variablePointSize_K, variablePointSize_DepthScale;
 
-out mediump vec4 frag_color;
+out highp vec4 frag_color;
 
 void main()
 {
-    mediump vec4 eye_position = mv_matrix * vec4(position, 1.0);
+    highp vec4 eye_position = mv_matrix * vec4(position, 1.0);
     gl_Position = p_matrix * eye_position;
     if (enableVariablePointSize!=0)
       gl_PointSize = vertexPointSize +
