+++ libs/opengl/shaders/textured-triangles.f.glsl
@@ -4,14 +4,14 @@ R"XXX(#version 300 es
 // Jose Luis Blanco Claraco (C) 2019-2020
 // Part of the MRPT project
 
-uniform mediump mat4 mv_matrix;
-uniform mediump vec4 light_diffuse, light_ambient;
-uniform mediump vec3 light_direction;
+uniform highp mat4 mv_matrix;
+uniform highp vec4 light_diffuse, light_ambient;
+uniform highp vec3 light_direction;
 uniform lowp int enableLight;  // 0 or 1
-uniform mediump sampler2D textureSampler;
+uniform highp sampler2D textureSampler;
 
-in mediump vec3 frag_position, frag_normal;
-in mediump vec2 frag_UV; // Interpolated values from the vertex shaders
+in highp vec3 frag_position, frag_normal;
+in highp vec2 frag_UV; // Interpolated values from the vertex shaders
 
 out highp vec4 color;
 
@@ -19,11 +19,11 @@ void main()
 {
     if (enableLight!=0)
     {
-        mediump vec3 mv_light_direction = light_direction;
-        mediump vec3 eye = normalize(frag_position);
-        mediump vec3 fn = normalize(frag_normal);
+        highp vec3 mv_light_direction = light_direction;
+        highp vec3 eye = normalize(frag_position);
+        highp vec3 fn = normalize(frag_normal);
 
-        mediump vec4 diffuse_factor = max(-dot(fn, mv_light_direction), 0.0) * light_diffuse;
+        highp vec4 diffuse_factor = max(-dot(fn, mv_light_direction), 0.0) * light_diffuse;
 
         color = texture( textureSampler, frag_UV ) * (diffuse_factor + light_ambient);
     }
